# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArthaWarga API Documentation",
        "description": "Dokumentasi resmi API ArthaWarga",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bop": {
            "get": {
                "tags": [
                    "BOP"
                ],
                "summary": "Ambil semua data pemasukan BOP",
                "operationId": "3f9550b5ef21c0ea0896b30407c01bd5",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bop/create": {
            "post": {
                "tags": [
                    "BOP"
                ],
                "summary": "Tambah pemasukan BOP",
                "operationId": "1fd8ceb57ed621c1d020bf0f46fd0fa8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "tgl",
                                    "nominal",
                                    "ket"
                                ],
                                "properties": {
                                    "tgl": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-25"
                                    },
                                    "nominal": {
                                        "type": "number",
                                        "example": 250000
                                    },
                                    "ket": {
                                        "type": "string",
                                        "example": "Pembelian ATK"
                                    },
                                    "bkt_nota": {
                                        "description": "Upload file nota (jpg, jpeg, png, pdf)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bop/delete/{id}": {
            "delete": {
                "tags": [
                    "BOP"
                ],
                "summary": "Hapus pemasukan BOP",
                "operationId": "6d2ea6f88af494898feb1deb43a6c425",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID data BOP",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bop/update/{id}": {
            "post": {
                "tags": [
                    "BOP"
                ],
                "summary": "Perbarui pemasukan BOP",
                "operationId": "372cf6797ddbf0e6d7be10f767594a06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID data BOP",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "tgl": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-01"
                                    },
                                    "nominal": {
                                        "type": "number",
                                        "example": 300000
                                    },
                                    "ket": {
                                        "type": "string",
                                        "example": "Perbaikan printer"
                                    },
                                    "bkt_nota": {
                                        "description": "Upload nota baru (opsional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}